Backend/restaurant_review/requirements.txt
asgiref==3.8.1
boto3==1.35.71
botocore==1.35.71
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Faker==37.1.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.1/libexec/wheel-0.45.1-py3-none-any.whl#sha256=da46333d5dcbde6e20cf7e2f8fff9e9ce76e8c94dc4afd6fb95fc4bc2745fb5e
